Factorize request set body into a new method request_set_body().
