+++ crates/nu-command/src/network/http/post.rs
@@ -1,4 +1,3 @@
-use crate::formats::value_to_json_value;
 use nu_engine::CallExt;
 use nu_protocol::ast::Call;
 use nu_protocol::engine::{Command, EngineState, Stack};
@@ -8,7 +7,7 @@ use nu_protocol::{
 
 use crate::network::http::client::{
     http_client, request_add_authorization_header, request_add_custom_headers,
-    request_handle_response, request_set_timeout,
+    request_handle_response, request_set_body, request_set_timeout,
 };
 
 #[derive(Clone)]
@@ -135,13 +134,6 @@ struct Arguments {
     content_length: Option<String>,
 }
 
-#[derive(PartialEq, Eq)]
-enum BodyType {
-    Json,
-    Form,
-    Unknown,
-}
-
 fn run_post(
     engine_state: &EngineState,
     stack: &mut Stack,
@@ -162,6 +154,7 @@ fn run_post(
     };
     helper(engine_state, stack, call, args)
 }
+
 // Helper function that actually goes to retrieve the resource from the url given
 // The Option<String> return a possible file extension which can be used in AutoConvert commands
 fn helper(
@@ -169,9 +162,17 @@ fn helper(
     stack: &mut Stack,
     call: &Call,
     args: Arguments,
-) -> std::result::Result<PipelineData, ShellError> {
+) -> Result<PipelineData, ShellError> {
     let url_value = args.path;
     let body = args.body;
+    let user = args.user.clone();
+    let password = args.password;
+    let timeout = args.timeout;
+    let headers = args.headers;
+    let content_type = args.content_type;
+    let content_length = args.content_length;
+    let raw = args.raw;
+
     let span = url_value.span()?;
     let requested_url = url_value.as_string()?;
     let url = match url::Url::parse(&requested_url) {
@@ -186,61 +187,10 @@ fn helper(
             ));
         }
     };
-    let user = args.user.clone();
-    let password = args.password;
-    let timeout = args.timeout;
-    let headers = args.headers;
-    let location = url;
-    let raw = args.raw;
 
-    let body_type = match &args.content_type {
-        Some(it) if it == "application/json" => BodyType::Json,
-        Some(it) if it == "application/x-www-form-urlencoded" => BodyType::Form,
-        _ => BodyType::Unknown,
-    };
+    let mut request = http_client(args.insecure.is_some()).post(url);
 
-    let mut request = http_client(args.insecure.is_some()).post(location);
-
-    // set the content-type header before using e.g., request.json
-    // because that will avoid duplicating the header value
-    if let Some(val) = args.content_type {
-        request = request.header("Content-Type", val);
-    }
-
-    match body {
-        Value::Binary { val, .. } => {
-            request = request.body(val);
-        }
-        Value::String { val, .. } => {
-            request = request.body(val);
-        }
-        Value::Record { .. } if body_type == BodyType::Json => {
-            let data = value_to_json_value(&body)?;
-            request = request.json(&data);
-        }
-        Value::Record { .. } if body_type == BodyType::Form => {
-            let data = value_to_json_value(&body)?;
-            request = request.form(&data);
-        }
-        Value::List { vals, .. } if body_type == BodyType::Form => {
-            if vals.len() % 2 != 0 {
-                return Err(ShellError::IOError("unsupported body input".into()));
-            }
-            let data = vals
-                .chunks(2)
-                .map(|it| Ok((it[0].as_string()?, it[1].as_string()?)))
-                .collect::<Result<Vec<(String, String)>, ShellError>>()?;
-            request = request.form(&data)
-        }
-        _ => {
-            return Err(ShellError::IOError("unsupported body input".into()));
-        }
-    };
-
-    if let Some(val) = args.content_length {
-        request = request.header("Content-Length", val);
-    }
-
+    request = request_set_body(content_type, content_length, body, request)?;
     request = request_set_timeout(timeout, request)?;
     request = request_add_authorization_header(user, password, request);
     request = request_add_custom_headers(headers, request)?;
